requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
selenium
webdriver-manager
python-dotenv